+++ game_patch/hud/hud.h
@@ -27,3 +27,4 @@ void toggle_chat_menu(ChatMenuType state);
 bool get_chat_menu_is_active();
 void hud_render_draw_chat_menu();
 void chat_menu_action_handler(rf::Key key);
+void build_local_player_spectators_strings();