+++ game_patch/hud/multi_scoreboard.h
@@ -3,3 +3,4 @@
 void multi_scoreboard_apply_patch();
 void multi_scoreboard_set_hidden(bool hidden);
 void multi_scoreboard_set_big(bool is_big);
+bool multi_scoreboard_is_visible();