+++ game_patch/misc/alpine_settings.cpp
@@ -220,6 +220,10 @@ bool alpine_player_settings_load(rf::Player* player)
         g_alpine_game_config.real_armor_values = std::stoi(settings["RealArmorValues"]);
         processed_keys.insert("RealArmorValues");
     }
+    if (settings.count("AlwaysShowSpectators")) {
+        g_alpine_game_config.always_show_spectators = std::stoi(settings["AlwaysShowSpectators"]);
+        processed_keys.insert("AlwaysShowSpectators");
+    }
     if (settings.count("ColorblindMode")) {
         int mode = std::stoi(settings["ColorblindMode"]);
         g_alpine_game_config.colorblind_mode = std::clamp(mode, 0, 3);
@@ -730,6 +734,7 @@ void alpine_player_settings_save(rf::Player* player)
     file << "SaveConsoleHistory=" << g_alpine_game_config.save_console_history << "\n";
     file << "AlpineBranding=" << g_alpine_game_config.af_branding << "\n";
     file << "RealArmorValues=" << g_alpine_game_config.real_armor_values << "\n";
+    file << "AlwaysShowSpectators=" << g_alpine_game_config.always_show_spectators << "\n";
     file << "ColorblindMode=" << g_alpine_game_config.colorblind_mode << "\n";
     file << "AutoswitchFireWait=" << g_alpine_game_config.suppress_autoswitch_fire_wait << "\n";
     file << "AlwaysAutoswitchEmpty=" << g_alpine_game_config.always_autoswitch_empty << "\n";