+++ game_patch/misc/alpine_settings.h
@@ -124,6 +124,7 @@ struct AlpineGameSettings
     bool static_bomb_code = false;
     bool entity_pain_sounds = true;
     bool real_armor_values = false;
+    bool always_show_spectators = false;
     int colorblind_mode = 0; // 0=off,1=protanopia,2=deuteranopia,3=tritanopia
     int suppress_autoswitch_alias = -1;
     bool always_autoswitch_empty = true;