+++ game_patch/misc/player.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <functional>
 #include <map>
 #include <optional>
@@ -8,6 +10,7 @@
 #include "../rf/math/matrix.h"
 #include "../rf/os/timestamp.h"
 #include "../purefaction/pf_packets.h"
+#include <unordered_set>
 
 // Forward declarations
 namespace rf
@@ -42,18 +45,22 @@ struct PlayerAdditionalData
     rf::TimestampRealtime idle_kick_timestamp;
     rf::Timestamp respawn_timer; // only used when configured in ADS
     uint8_t damage_handicap = 0; // percentile
+    std::optional<rf::Player*> spectatee{};
 };
 
 inline rf::Timestamp g_respawn_timer_local;
 inline bool g_spawned_in_current_level = false; // relevant if force respawn is on
 inline bool g_local_queued_delayed_spawn = false;
+inline std::unordered_set<rf::Player*> g_local_player_spectators{};
+inline std::string g_local_player_spectators_spawned_string{};
+inline std::string g_local_player_spectators_unspawned_string{};
 
 std::string build_local_spawn_string(bool can_respawn);
 void set_local_spawn_delay(bool can_respawn, bool force_respawn, int spawn_delay);
 void reset_local_delayed_spawn();
 void find_player(const StringMatcher& query, std::function<void(rf::Player*)> consumer);
 void reset_player_additional_data(const rf::Player* player);
-PlayerAdditionalData& get_player_additional_data(rf::Player* player);
+PlayerAdditionalData& get_player_additional_data(const rf::Player* player);
 void play_local_hit_sound(bool died);
 void handle_chat_message_sound(std::string message);
 bool is_player_minimum_af_client_version(rf::Player* player, int version_major, int version_minor);